evals/requirements.txt
pyyaml==6.0.1
openai==1.12.0
anthropic==0.18.1
langfuse==2.28.3
httpx==0.26.0
python-dotenv==1.0.1
